Fix cannot send test message (security update) (#293)
Full details: https://github.com/chiefonboarding/ChiefOnboarding/pull/293
There was a bug which allowed the user's 2FA secret token to be empty. This was only the case if the user was created through the initial setup wizard which has been introduced on March 28. Running the migration of this commit will fix that, but might not allow a user to pass 2FA (as a new token will be created). Anyone already using ChiefOnboarding before that release is not affected.